Find best split automatically.
Between two consecutive note shards, move rests in such a fashion
as make the split feel the most natural.
For example, if possible, split on a completed measurement. A half
measurement is more natural than a quarter, wich in turn is more
natural than an eighth. Weird numbers like 16*3 are penalized more
heavily.